requirements.txt
requests==2.31.0
playwright==1.40.0
PyPDF2==3.0.1
python-dotenv==1.0.0
flask==3.0.0
gunicorn==21.2.0
pydantic==2.5.0
asyncio==3.4.3
